Rename services/funding to services/gas-station
'Funding' reads as financial funding; the folder holds gas provisioning
(GasStation + WalletActivator).
