+++ src/integrations/deposits/pull-deposit/plugin.ts
@@ -16,7 +16,7 @@ import {
 } from "@owneraio/finp2p-nodejs-skeleton-adapter";
 import { FinP2PClient } from "@owneraio/finp2p-client";
 import { CustodyProvider, CustodyWallet } from "../../../services/custody";
-import { GasStation } from "../../../services/funding";
+import { GasStation } from "../../../services/gas-station";
 import { AssetStore } from "../../../services/accounts";
 import { DepositTargetResolver } from "../types";
 import { ApprovalWatcher } from "./approval-watcher";
